requirements.txt
annotated-types
anyio
APScheduler
bcrypt
black
boto3
botocore
certifi
cffi
charset-normalizer
click
cryptography
dnspython
ecdsa
email-validator
et_xmlfile
fastapi
flake8
greenlet
h11
httpcore
httpx
idna
iniconfig
isort
jmespath
jq
librt
markdown-it-py
mccabe
mdurl
motor
mypy
mypy_extensions
numpy
oauthlib
openpyxl
packaging
pandas
passlib
pathspec
pillow
platformdirs
pluggy
pyasn1
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
PyMySQL
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
python-telegram-bot
pytokens
pytz
qrcode
reportlab
requests
requests-oauthlib
rich
rsa
s3transfer
s5cmd
shellingham
six
SQLAlchemy
starlette
typer
typing-inspection
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
watchfiles
